Adds View menu containing Toggle Fullscreen and Reload
