+++ src/menu/base_menu_template.js
@@ -1,12 +1,13 @@
 import { appMenuTemplate } from './app_menu_template';
 import { fileMenuTemplate } from './file_menu_template';
 import { editMenuTemplate } from './edit_menu_template';
-import { windowMenuTemplate } from './window_menu_template';
 import { settingsMenu } from './settings_menu_template';
+import { viewMenuTemplate } from './view_menu_template';
+import { windowMenuTemplate } from './window_menu_template';
 import { IS_MAC } from '../constants';
 
 
-const baseMenuTemplate = [editMenuTemplate, windowMenuTemplate];
+const baseMenuTemplate = [editMenuTemplate, viewMenuTemplate, windowMenuTemplate];
 
 if (IS_MAC) {
     baseMenuTemplate.unshift(appMenuTemplate);
